requirements.txt
matplotlib
google-cloud-speech
boto3
deepspeech==0.6.1 # deepspeech==0.5.0
editdistance
numpy
requests
soundfile
beautifulsoup4
num2words
azure-cognitiveservices-speech
watson-developer-cloud
ibm-watson
lxml
